+++ src/compression/mod.rs
@@ -7,10 +7,11 @@ pub enum Compression {
     None,
     #[cfg(feature = "lz4")]
     Lz4,
+    /// High compression levels are useful in networks with low bandwidth.
+    /// Affects only INSERTs, because others are compressed by the server.
+    /// Possible levels: [1, 12]. Recommended level range: [4, 9].
     #[cfg(feature = "lz4")]
     Lz4Hc(i32),
-    #[cfg(feature = "lz4")]
-    Lz4Fast(i32),
 }
 
 impl Default for Compression {
@@ -28,13 +29,7 @@ impl Default for Compression {
 }
 
 impl Compression {
-    #[cfg(feature = "lz4")]
     pub(crate) fn is_lz4(&self) -> bool {
         *self != Compression::None
-    }
-
-    #[cfg(not(feature = "lz4"))]
-    pub(crate) fn is_lz4(&self) -> bool {
-        false
     }
 }